requirements.txt
# Core dependencies
pydantic>=2.0.0
pyyaml>=6.0.0

# LLM API
anthropic>=0.25.0

# PDF Processing
PyMuPDF>=1.23.0
pdfplumber>=0.10.0

# Reporting
jinja2>=3.1.0
markdown>=3.5.0

# CLI
click>=8.1.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0
